requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
scikit-learn
PyMuPDF
llama-cpp-python
langchain
streamlit
